Extract DataFileRecord
Share DataFileRecord functionality across the Iceberg tests
