+++ plugin/trino-iceberg/src/test/java/io/trino/plugin/iceberg/TestIcebergOrcMetricsCollection.java
@@ -13,7 +13,6 @@
  */
 package io.trino.plugin.iceberg;
 
-import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import io.trino.Session;
 import io.trino.plugin.hive.HdfsConfig;
@@ -49,7 +48,7 @@ import java.util.Optional;
 import static io.trino.SystemSessionProperties.MAX_DRIVERS_PER_TASK;
 import static io.trino.SystemSessionProperties.TASK_CONCURRENCY;
 import static io.trino.SystemSessionProperties.TASK_WRITER_COUNT;
-import static io.trino.plugin.iceberg.TestIcebergOrcMetricsCollection.DataFileRecord.toDataFileRecord;
+import static io.trino.plugin.iceberg.DataFileRecord.toDataFileRecord;
 import static io.trino.testing.TestingSession.testSessionBuilder;
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertNull;
@@ -408,118 +407,5 @@ public class TestIcebergOrcMetricsCollection
         assertQuery("SELECT max(_timestamp) FROM test_timestamp", "VALUES '2021-01-31 00:00:00.333333'");
 
         assertUpdate("DROP TABLE test_timestamp");
-    }
-
-    public static class DataFileRecord
-    {
-        private final int content;
-        private final String filePath;
-        private final String fileFormat;
-        private final long recordCount;
-        private final long fileSizeInBytes;
-        private final Map<Integer, Long> columnSizes;
-        private final Map<Integer, Long> valueCounts;
-        private final Map<Integer, Long> nullValueCounts;
-        private final Map<Integer, Long> nanValueCounts;
-        private final Map<Integer, String> lowerBounds;
-        private final Map<Integer, String> upperBounds;
-
-        public static DataFileRecord toDataFileRecord(MaterializedRow row)
-        {
-            assertEquals(row.getFieldCount(), 14);
-            return new DataFileRecord(
-                    (int) row.getField(0),
-                    (String) row.getField(1),
-                    (String) row.getField(2),
-                    (long) row.getField(3),
-                    (long) row.getField(4),
-                    row.getField(5) != null ? ImmutableMap.copyOf((Map<Integer, Long>) row.getField(5)) : null,
-                    row.getField(6) != null ? ImmutableMap.copyOf((Map<Integer, Long>) row.getField(6)) : null,
-                    row.getField(7) != null ? ImmutableMap.copyOf((Map<Integer, Long>) row.getField(7)) : null,
-                    row.getField(8) != null ? ImmutableMap.copyOf((Map<Integer, Long>) row.getField(8)) : null,
-                    row.getField(9) != null ? ImmutableMap.copyOf((Map<Integer, String>) row.getField(9)) : null,
-                    row.getField(10) != null ? ImmutableMap.copyOf((Map<Integer, String>) row.getField(10)) : null);
-        }
-
-        private DataFileRecord(
-                int content,
-                String filePath,
-                String fileFormat,
-                long recordCount,
-                long fileSizeInBytes,
-                Map<Integer, Long> columnSizes,
-                Map<Integer, Long> valueCounts,
-                Map<Integer, Long> nullValueCounts,
-                Map<Integer, Long> nanValueCounts,
-                Map<Integer, String> lowerBounds,
-                Map<Integer, String> upperBounds)
-        {
-            this.content = content;
-            this.filePath = filePath;
-            this.fileFormat = fileFormat;
-            this.recordCount = recordCount;
-            this.fileSizeInBytes = fileSizeInBytes;
-            this.columnSizes = columnSizes;
-            this.valueCounts = valueCounts;
-            this.nullValueCounts = nullValueCounts;
-            this.nanValueCounts = nanValueCounts;
-            this.lowerBounds = lowerBounds;
-            this.upperBounds = upperBounds;
-        }
-
-        public int getContent()
-        {
-            return content;
-        }
-
-        public String getFilePath()
-        {
-            return filePath;
-        }
-
-        public String getFileFormat()
-        {
-            return fileFormat;
-        }
-
-        public long getRecordCount()
-        {
-            return recordCount;
-        }
-
-        public long getFileSizeInBytes()
-        {
-            return fileSizeInBytes;
-        }
-
-        public Map<Integer, Long> getColumnSizes()
-        {
-            return columnSizes;
-        }
-
-        public Map<Integer, Long> getValueCounts()
-        {
-            return valueCounts;
-        }
-
-        public Map<Integer, Long> getNullValueCounts()
-        {
-            return nullValueCounts;
-        }
-
-        public Map<Integer, Long> getNanValueCounts()
-        {
-            return nanValueCounts;
-        }
-
-        public Map<Integer, String> getLowerBounds()
-        {
-            return lowerBounds;
-        }
-
-        public Map<Integer, String> getUpperBounds()
-        {
-            return upperBounds;
-        }
     }
 }
